add dm, group, channel load history routes
