+++ rest/channels.go
@@ -4,6 +4,8 @@ import (
 	"bytes"
 	"fmt"
 	"net/url"
+	"strconv"
+	"time"
 
 	"github.com/RocketChat/Rocket.Chat.Go.SDK/models"
 )
@@ -109,4 +111,28 @@ func (c *Client) GetGroupInfo(channel *models.Channel) (*models.Channel, error) 
 		}
 	}
 	return &response.Group, nil
+}
+
+// Get messages from a channel. The channel id has to be not nil. Optionally a
+// count can be specified to limit the size of the returned messages.
+//
+// https://rocket.chat/docs/developer-guides/rest-api/channels/history
+func (c *Client) ChannelHistory(channel *models.Channel, inclusive bool, fromDate time.Time, page *models.Pagination) ([]models.Message, error) {
+	params := url.Values{
+		"roomId": []string{channel.ID},
+	}
+
+	if page != nil {
+		params.Add("count", strconv.Itoa(page.Count))
+	}
+
+	params.Add("inclusive", "true")
+	params.Add("oldest", fromDate.String())
+
+	response := new(MessagesResponse)
+	if err := c.Get("channels.history", params, response); err != nil {
+		return nil, err
+	}
+
+	return response.Messages, nil
 }
